moved json representation of the NetworkConfig next to the NetworkConfig
