+++ chain/network-primitives/src/lib.rs
@@ -1,4 +1,5 @@
 mod blacklist;
 pub(crate) mod config;
+pub(crate) mod config_json;
 mod network_protocol;
 pub mod types;
