+++ chain/network-primitives/src/types.rs
@@ -36,8 +36,8 @@ pub use crate::network_protocol::{
 };
 
 pub use crate::blacklist::Blacklist;
-pub use crate::config::{NetworkConfig,ValidatorConfig,ValidatorEndpoints};
-
+pub use crate::config::{NetworkConfig, ValidatorConfig, ValidatorEndpoints};
+pub use crate::config_json::Config as ConfigJSON;
 pub use crate::network_protocol::edge::{Edge, EdgeState, PartialEdgeInfo, SimpleEdge};
 
 /// Number of hops a message is allowed to travel before being dropped.
