+++ integration-tests/src/tests/network/peer_manager.rs
@@ -99,8 +99,7 @@ fn repeated_announce_accounts() {
                 _ => NetworkViewClientResponses::NoResponse,
             }
         })));
-        let routing_table_addr =
-            start_routing_table_actor(config.node_id(), store.clone());
+        let routing_table_addr = start_routing_table_actor(config.node_id(), store.clone());
         let pm = PeerManagerActor::new(
             store,
             config,
